requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
backports.tarfile==1.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.4
datasets==3.0.1
dill==0.3.6
docutils==0.21.2
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.25.1
idna==3.8
importlib_metadata==8.5.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
joblib==1.4.2
keyring==25.4.1
lantern-client==0.0.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.14
networkx==3.3
nh3==0.2.18
nltk==3.9.1
numpy==1.24.3
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pkginfo==1.10.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
readme_renderer==44.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.18.0
rfc3986==2.0.0
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
sentencepiece==0.2.0
six==1.16.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.2.2
torchvision==0.17.2
towhee==1.1.3
towhee.models==1.1.3
tqdm==4.66.5
transformers==4.45.1
twine==5.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
Werkzeug==3.0.4
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
